Now you can add multiple pieces (But some things are still to fix)
Multiple pieces are now possible thanks to coroutines.
But for now it's jumping by two and shows only one color of it.
Need to Fix AddPlates method.

Probably it would be good to find an alternative for searching for movePlateCount on Update because after the first State it would be not necessary. Maybe make it as an instance of another object to kill after the first state?